# lib/api-zod/src/generated/types/publishObjectInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublishObjectInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  objectPath: string;
}